Improve cover-pkg calculation methods
This update changes the way that totals are calculated by parsing the
go.cov output directly.  When not using cover-pkg the results should be
the same.  For cover-pkg the results should now be more inline with what
would typically be expected (i.e if pkg1 is completely covered by three
different packages, then the result for pkg1 shoudl be 100%). See the
readme for more info

Also adds a new `ignore-pattern` option so that generated code can be
excluded from the coverage calculations and makes some output formatting
improvements.